editoast: isolate search CLI in its own module
